Fixes floating point gpt conditioning lenghts.

+++ TTS/tts/models/xtts.py
@@ -271,8 +271,8 @@ class Xtts(BaseTTS):
             audio = audio[:, : 22050 * length]
         if self.args.gpt_use_perceiver_resampler:
             style_embs = []
-            for i in range(0, audio.shape[1], 22050 * chunk_length):
-                audio_chunk = audio[:, i : i + 22050 * chunk_length]
+            for i in range(0, audio.shape[1], int(22050 * chunk_length)):
+                audio_chunk = audio[:, i : i + int(22050 * chunk_length)]
 
                 # if the chunk is too short ignore it 
                 if audio_chunk.size(-1) < 22050 * 0.33:
